fix: only track fetch quotes event one time (#10234)
